Moved CardExtension to its own package.
