+++ kansha/card/comp.py
@@ -17,7 +17,7 @@ from kansha.toolbox import overlay
 from kansha.user import usermanager
 from kansha import title
 from kansha import exceptions, notifications
-from kansha.services.components_repository import CardExtension
+from kansha.cardextension import CardExtension
 
 from .models import DataCard
