+++ kansha/cardextension/__init__.py
@@ -1,0 +1,11 @@
+#--
+# Copyright (c) 2012-2014 Net-ng.
+# All rights reserved.
+#
+# This software is licensed under the BSD License, as described in
+# the file LICENSE.txt, which you should have received as part of
+# this distribution.
+#--
+
+from . import view
+from .comp import CardExtension
