+++ kansha/checklist/comp.py
@@ -14,7 +14,7 @@ from nagare import component, database, i18n, security
 
 from kansha import notifications
 from kansha import title
-from kansha.services.components_repository import CardExtension
+from kansha.cardextension import CardExtension
 
 from .models import DataChecklist, DataChecklistItem
