+++ kansha/comment/comp.py
@@ -15,7 +15,7 @@ from nagare import component, security
 
 from kansha.user import usermanager
 from kansha import notifications, validator
-from kansha.services.components_repository import CardExtension
+from kansha.cardextension import CardExtension
 
 from .models import DataComment
