+++ kansha/description/comp.py
@@ -8,7 +8,7 @@
 # this distribution.
 #--
 from kansha import validator
-from kansha.services.components_repository import CardExtension
+from kansha.cardextension import CardExtension
 
 
 class CardDescription(CardExtension):
