+++ kansha/due_date/comp.py
@@ -12,7 +12,7 @@ from datetime import date
 from nagare import component
 
 from kansha.toolbox import calendar_widget
-from kansha.services.components_repository import CardExtension
+from kansha.cardextension import CardExtension
 
 
 class DueDate(CardExtension):
