+++ kansha/gallery/comp.py
@@ -19,7 +19,7 @@ from kansha import notifications
 from kansha.toolbox import overlay
 from kansha.user import usermanager
 from kansha.authentication.database import validators
-from kansha.services.components_repository import CardExtension
+from kansha.cardextension import CardExtension
 
 from .models import DataGallery, DataAsset
