+++ kansha/label/comp.py
@@ -13,7 +13,7 @@ import random
 from nagare import component, var
 
 from kansha.toolbox import overlay
-from kansha.services.components_repository import CardExtension
+from kansha.cardextension import CardExtension
 
 from .models import DataLabel
